rename the data transfer class to data management class for saving and loading functionality and add the remaining models

+++ src/InventBox.Test/Utils/FileUtils.cs
@@ -4,8 +4,11 @@ namespace InventBox.Test.Utils;
 
 public class FileUtils
 {
-    public static void CleanUpFile(string path)
+    public static void CleanUpFile(string[] paths)
     {
-        File.Delete(path);
+        foreach (var path in paths) {
+            if (File.Exists(path))
+                File.Delete(path);
+        }
     }
 }
